test refactoring to consistently use a single Blocker (as much as possible)
