+++ tests/src/test/scala/com/dwolla/security/crypto/CryptoAlgSpec.scala
@@ -4,36 +4,32 @@ import cats.effect._
 import cats.effect.testing.scalatest.CatsResourceIO
 import cats.syntax.all._
 import com.dwolla.testutils._
-import eu.timepit.refined.api.Refined
-import eu.timepit.refined.auto._
-import eu.timepit.refined.numeric.Positive
-import eu.timepit.refined.scalacheck.all._
 import fs2._
 import org.bouncycastle.bcpg._
 import org.bouncycastle.openpgp._
-import org.bouncycastle.openpgp.operator.jcajce.{JcaPGPContentSignerBuilder, JcaPGPDigestCalculatorProviderBuilder, JcePBESecretKeyEncryptorBuilder}
 import org.scalacheck.Arbitrary._
 import org.scalacheck._
 import org.scalatest.flatspec._
 import org.typelevel.log4cats.Logger
-import org.scalacheck.cats.implicits._
 
 import java.io.ByteArrayOutputStream
-import scala.jdk.CollectionConverters._
 
-class CryptoAlgSpec1
+class CryptoAlgSpec
   extends FixtureAsyncFlatSpec
-    with CatsResourceIO[CryptoAlg[IO]]
+    with CatsResourceIO[(Blocker, CryptoAlg[IO])]
     with CryptoArbitraries
     with Fs2PgpSpec {
 
-  override def resource: Resource[IO, CryptoAlg[IO]] = Blocker[IO].flatMap(CryptoAlg[IO](_))
+  override def resource: Resource[IO, (Blocker, CryptoAlg[IO])] = Blocker[IO].mproduct(CryptoAlg[IO](_))
 
   private implicit val noOpLogger: Logger[IO] = NoOpLogger[IO]()
 
   behavior of "CryptoAlg"
 
-  it should "round trip the plaintext" in { crypto =>
+  it should "round trip the plaintext" in { tuple =>
+    val (blocker, crypto) = tuple
+    implicit val arbKeyPair: Arbitrary[Resource[IO, PGPKeyPair]] = arbWeakKeyPair(blocker)
+
     forAll(MinSuccessful(1)) { (keyPairR: Resource[IO, PGPKeyPair],
                                 bytesG: Stream[Pure, Byte],
                                 encryptionChunkSize: ChunkSize,
@@ -54,21 +50,11 @@ class CryptoAlgSpec1
       }
     }
   }
-}
 
-class CryptoAlgSpec2
-  extends FixtureAsyncFlatSpec
-    with CatsResourceIO[CryptoAlg[IO]]
-    with CryptoArbitraries
-    with Fs2PgpSpec {
+  it should "maintain chunk size throughout pipeline" in { tuple =>
+    val (blocker, crypto) = tuple
+    implicit val arbKeyPair: Arbitrary[Resource[IO, PGPKeyPair]] = arbWeakKeyPair(blocker)
 
-  override def resource: Resource[IO, CryptoAlg[IO]] = Blocker[IO].flatMap(CryptoAlg[IO](_))
-
-  private implicit val noOpLogger: Logger[IO] = NoOpLogger[IO]()
-
-  behavior of "CryptoAlg"
-
-  it should "maintain chunk size throughout pipeline" in { crypto =>
     forAll(MinSuccessful(1)) { (keyPairR: Resource[IO, PGPKeyPair],
                                 encryptionChunkSize: ChunkSize) =>
       // since the cryptotext is compressed, we need to generate at least 10x the chunk size to
@@ -91,21 +77,11 @@ class CryptoAlgSpec2
       }
     }
   }
-}
 
-class CryptoAlgSpec3
-  extends FixtureAsyncFlatSpec
-    with CatsResourceIO[CryptoAlg[IO]]
-    with CryptoArbitraries
-    with Fs2PgpSpec {
+  it should "support armoring a PGP value" in { tuple =>
+    val (blocker, crypto) = tuple
+    implicit val arbKeyPair: Arbitrary[Resource[IO, PGPKeyPair]] = arbWeakKeyPair(blocker)
 
-  override def resource: Resource[IO, CryptoAlg[IO]] = Blocker[IO].flatMap(CryptoAlg[IO](_))
-
-  private implicit val noOpLogger: Logger[IO] = NoOpLogger[IO]()
-
-  behavior of "CryptoAlg"
-
-  it should "support armoring a PGP value" in { crypto =>
     forAll(arbPgpBytes[IO].arbitrary, MinSuccessful(1)) { (bytesR: Resource[IO, Array[Byte]]) =>
       for {
         blocker <- Blocker[IO]
@@ -123,37 +99,10 @@ class CryptoAlgSpec3
       }
     }
   }
-}
-
-class CryptoAlgSpec4
-  extends FixtureAsyncFlatSpec
-    with CatsResourceIO[(Blocker, CryptoAlg[IO])]
-    with CryptoArbitraries
-    with Fs2PgpSpec {
-
-  override def resource: Resource[IO, (Blocker, CryptoAlg[IO])] = Blocker[IO].mproduct(CryptoAlg[IO](_))
-
-  private implicit val noOpLogger: Logger[IO] = NoOpLogger[IO]()
-
-  behavior of "CryptoAlg"
-
-  def genPGPSecretKeyRingCollection[F[_] : Sync : ContextShift : Clock](blocker: Blocker,
-                                                                        passphrase: Array[Char]): Gen[Resource[F, PGPSecretKeyRingCollection]] =
-    (arbitrary[Resource[F, PGPKeyPair]], arbitrary[String]).mapN { (keyPairR, keyRingId) =>
-      for {
-        kp <- keyPairR
-        generator <- Resource.eval(pgpKeyRingGenerator[F](blocker)(keyRingId, kp, passphrase))
-      } yield new PGPSecretKeyRingCollection(List(generator.generateSecretKeyRing()).asJava)
-    }
-
-  private def keysIn[F[_] : Sync](collection: PGPSecretKeyRingCollection): Stream[F, PGPSecretKey] =
-    for {
-      ring <- Stream.fromIterator[F](collection.iterator().asScala)
-      key <- Stream.fromIterator[F](ring.iterator().asScala)
-    } yield key
 
   it should "round trip the plaintext using a key ring collection" in { tuple =>
     val (blocker, crypto) = tuple
+    implicit val arbKeyPair: Arbitrary[Resource[IO, PGPKeyPair]] = arbWeakKeyPair(blocker)
 
     forAll(MinSuccessful(1)) { (passphrase: Array[Char],
                                 bytesG: Stream[Pure, Byte],
@@ -178,22 +127,10 @@ class CryptoAlgSpec4
       }
     }
   }
-}
-
-class CryptoAlgSpec5
-  extends FixtureAsyncFlatSpec
-    with CatsResourceIO[(Blocker, CryptoAlg[IO])]
-    with CryptoArbitraries
-    with Fs2PgpSpec {
-
-  override def resource: Resource[IO, (Blocker, CryptoAlg[IO])] = Blocker[IO].mproduct(CryptoAlg[IO](_))
-
-  private implicit val noOpLogger: Logger[IO] = NoOpLogger[IO]()
-
-  behavior of "CryptoAlg"
 
   it should "round trip the plaintext using a key ring" in { tuple =>
     val (blocker, crypto) = tuple
+    implicit val arbKeyPair: Arbitrary[Resource[IO, PGPKeyPair]] = arbWeakKeyPair(blocker)
 
     forAll(MinSuccessful(1)) { (keyPairR: Resource[IO, PGPKeyPair],
                                 bytesG: Stream[Pure, Byte],
@@ -219,49 +156,4 @@ class CryptoAlgSpec5
     }
   }
 
-}
-
-trait CryptoArbitraries { self: PgpArbitraries =>
-  implicit def arbKeyPair[F[_] : Sync : ContextShift : Clock]: Arbitrary[Resource[F, PGPKeyPair]] = arbWeakKeyPair[F]
-
-  def genNBytesBetween(min: Int, max: Int): Gen[Stream[Pure, Byte]] =
-    for {
-      count <- Gen.chooseNum(min, Math.max(min, max))
-      moreBytes <- Gen.listOfN(count, arbitrary[Byte])
-    } yield Stream.emits(moreBytes)
-
-  implicit val arbBytes: Arbitrary[Stream[Pure, Byte]] = Arbitrary {
-    genNBytesBetween(1 << 10, 1 << 20) // 1KB to 1MB
-  }
-
-  def arbPgpBytes[F[_] : Sync : ContextShift : Clock]: Arbitrary[Resource[F, Array[Byte]]] = Arbitrary {
-    for {
-      keyPair <- arbitrary[Resource[F, PGPKeyPair]]
-      bytes <- Gen.oneOf[Resource[F, Array[Byte]]](keyPair.map(_.getPublicKey.getEncoded), keyPair.map(_.getPrivateKey.getPrivateKeyDataPacket.getEncoded))
-    } yield bytes
-  }
-
-  implicit val arbChunkSize: Arbitrary[ChunkSize] = Arbitrary {
-    chooseRefinedNum[Refined, Int, Positive](1024, 4096).map(tagChunkSize)
-  }
-
-  def pgpKeyRingGenerator[F[_] : Sync : ContextShift](blocker: Blocker)
-                                                     (keyRingId: String,
-                                                      keyPair: PGPKeyPair,
-                                                      passphrase: Array[Char]): F[PGPKeyRingGenerator] =
-    blocker.delay {
-      val pgpContentSignerBuilder = new JcaPGPContentSignerBuilder(keyPair.getPublicKey.getAlgorithm, HashAlgorithmTags.SHA1)
-      val dc = new JcaPGPDigestCalculatorProviderBuilder().build().get(HashAlgorithmTags.SHA1)
-      val keyEncryptor = new JcePBESecretKeyEncryptorBuilder(SymmetricKeyAlgorithmTags.CAST5).build(passphrase)
-
-      new PGPKeyRingGenerator(PGPSignature.POSITIVE_CERTIFICATION,
-        keyPair,
-        keyRingId,
-        dc,
-        null,
-        null,
-        pgpContentSignerBuilder,
-        keyEncryptor
-      )
-    }
 }
