+++ tests/src/test/scala/com/dwolla/security/crypto/MultiThreadedReadOutputStreamSpec.scala
@@ -1,33 +1,39 @@
 package com.dwolla.security.crypto
 
 import cats.effect._
-import com.dwolla.testutils.NoOpLogger
+import cats.effect.testing.scalatest.CatsResourceIO
+import com.dwolla.testutils._
 import fs2._
 import org.bouncycastle.openpgp.PGPKeyPair
-import org.scalatest.flatspec.AsyncFlatSpec
+import org.scalacheck.Arbitrary
+import org.scalatest.flatspec.FixtureAsyncFlatSpec
 import org.typelevel.log4cats.Logger
 
 import java.util.concurrent.TimeUnit.SECONDS
 import java.util.concurrent.{SynchronousQueue, ThreadPoolExecutor}
 
 class MultiThreadedReadOutputStreamSpec
-  extends AsyncFlatSpec
+  extends FixtureAsyncFlatSpec
+    with CatsResourceIO[Blocker]
     with CryptoArbitraries
     with Fs2PgpSpec {
+
+  override def resource: Resource[IO, Blocker] = Blocker.fromExecutorService[IO](IO {
+    new ThreadPoolExecutor(0, Int.MaxValue, 0L, SECONDS, new SynchronousQueue[Runnable])
+  })
 
   private implicit val noOpLogger: Logger[IO] = NoOpLogger[IO]()
 
   behavior of "CryptoAlg"
 
-  it should "round trip the plaintext with a pathological ThreadPool" in {
+  it should "round trip the plaintext with a pathological ThreadPool" in { blocker =>
+    implicit val arbKeyPair: Arbitrary[Resource[IO, PGPKeyPair]] = arbWeakKeyPair(blocker)
+
     forAll(MinSuccessful(1)) { (keyPairR: Resource[IO, PGPKeyPair],
                                 bytes: Stream[Pure, Byte],
                                 encryptionChunkSize: ChunkSize,
                                 decryptionChunkSize: ChunkSize) =>
       for {
-        blocker <- Blocker.fromExecutorService[IO](IO {
-          new ThreadPoolExecutor(0, Int.MaxValue, 0L, SECONDS, new SynchronousQueue[Runnable])
-        })
         crypto <- CryptoAlg[IO](blocker)
         keyPair <- keyPairR
         roundTrip <- bytes
